Move support JS files inline with the test
